Get Jest Working
Ran `npm install --save-dev jest babel-jest ts-jest chai chai-jest identity-obj-proxy`

jest is a unit test runner from Facebook. You probably already know this, because you’re reading this.

ts-jest is a transform for jest which compiles typescript files. If you’re using babel to compile your typescript files, you can skip this.

babel-jest is like ts-jest, but uses babel to transform files - handy if you have a project with some mixed typescript and javascript.

chai is an assertion library. Jest already comes with an expect built in, but if you’re coming from mocha you probably already use chai, and it’s somewhat more expressive and has a lot of plugins available.

chai-jest is a plugin for chai which has supports jest mocks. It basically re-implements a bunch of the jest-specific asserts on the built-in expect object. See the documentation for a list of assertions available.

identity-obj-proxy is a handy library for cases where we import files like CSS modules. If you import styles from 'styles.css', then we can configure jest to import ‘identity-obj-proxy’ for *.css, and then when you do styles.container, it will resolve to “container” instead of throwing an exception.